Improve flow of types for translations and pinger schema
